# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8005/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/company/batches": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "summary": "List batches",
                "operationId": "b6c9a2de2368615cd8339af58bf4b52d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of batches"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/company/batch/{id}": {
            "get": {
                "tags": [
                    "Batches"
                ],
                "summary": "Get batch by ID",
                "operationId": "3947a469903c69ff19cf06574b27de6e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch found"
                    },
                    "404": {
                        "description": "Batch not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Batches"
                ],
                "summary": "Update a batch",
                "operationId": "acf61e313782f666bf0a4117421157e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Batch updated"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Batch not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Batches"
                ],
                "summary": "Delete a batch",
                "operationId": "7e8f0041e5e85f3d8d860d9015b3236a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Batch deleted"
                    },
                    "404": {
                        "description": "Batch not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/company/batch": {
            "post": {
                "tags": [
                    "Batches"
                ],
                "summary": "Create a batch",
                "operationId": "cf3100647d60105ef11ac6676af20e1a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tankId",
                                    "name",
                                    "entryDate",
                                    "initialQuantity",
                                    "species",
                                    "cultivation"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Batch created"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}